Refactor experiments to remove cycle in dependencies
